safari-534.53-branch/LayoutTests/fast/regex/cross-frame-callable-expected.txt

PASS re('a') threw exception TypeError: '/a/' is not a function (evaluating 're('a')').
PASS re.exec('a') is ['a']
PASS successfullyParsed is true

TEST COMPLETE
DONE
